### 2018-09-21 #968 @lbenet
@arfon I didn't know that the paper has to be in master. The paper can be found at https://github.com/JuliaDiff/TaylorSeries.jl/pull/181; shall I merge it to master?

Cc: @dpsanders, coauthor of the package and paper.
