### 2018-09-22 #968 @jedbrown
@lbenet Whedon takes the paper from 'master' (unless run manually). If it isn't overly disruptive to you, merging will make the automation do its job. The paper should be part of the repository that is eventually tagged (after acceptance) anyway.
