### 2018-09-22 3434e51
JOSS paper (#181)
* Add paper.md and paper.bib for an eventual submission to JOSS

* Small changes in paper.md

* Add a reference to TaylorModels.jl and fix typos

* Include David's suggestions

* Another tiny correction
