+++ paper/paper.bib
@@ -1,0 +1,83 @@
+@misc{TaylorSeries,
+  author       = {Benet, Luis and Sanders, David P.},
+  title        = {A julia package for Taylor expansions in one or
+    more independent variables},
+  howpublished = {\url{https://github.com/JuliaDiff/TaylorSeries.jl}}
+}
+
+@article{julia,
+	author  = {Jeff Bezanson and Alan Edelman and Stefan Karpinski and Viral B. Shah},
+	title   = {Julia: A Fresh Approach to Numerical Computing},
+	journal = {SIAM Review},
+	volume  = {59},
+	number  = {1},
+	pages   = {65-98},
+	year    = {2017},
+	doi     = {10.1137/141000671},
+}
+
+@book{Tucker:ValidatedNumerics,
+ author    = {Tucker, Warwick},
+ title     = {Validated Numerics: A Short Introduction to Rigorous Computations},
+ year      = {2011},
+ isbn      = {0691147817, 9780691147819},
+ publisher = {Princeton University Press},
+ address   = {Princeton, NJ, USA},
+}
+
+@book{HaroEtAl:ParameterizMeth,
+  author    = {Haro, Àlex and Canadell, Marta and Figueras, Jordi-Lluis
+    and Luque, Alejandro and Mondelo, Josep Maria},
+  title     = {The parameterization method for invariant manifolds: from
+    rigorous results to effective computations},
+  publisher = {Springer International Publishing},
+  address   = {Switzerland},
+  series    = {Applied mathematical sciences},
+  year      = {2016},
+  url       = {https://cds.cern.ch/record/2151752},
+}
+
+@article{MPFR,
+ author    = {Fousse, Laurent and Hanrot, Guillaume and Lef\`{e}vre, Vincent and P{\'e}lissier, Patrick and Zimmermann, Paul},
+ title     = {MPFR: A Multiple-precision Binary Floating-point Library with Correct Rounding},
+ journal   = {ACM Trans. Math. Softw.},
+ issue_date = {June 2007},
+ volume    = {33},
+ number    = {2},
+ month     = jun,
+ year      = {2007},
+ issn      = {0098-3500},
+ articleno = {13},
+ url       = {http://doi.acm.org/10.1145/1236463.1236468},
+ doi       = {10.1145/1236463.1236468},
+ acmid     = {1236468},
+ publisher = {ACM},
+ address   = {New York, NY, USA},
+ keywords  = {IEEE 754 standard, Multiple-precision arithmetic, correct rounding, elementary function, floating-point arithmetic, portable software},
+}
+
+@misc{ValidatedNumerics.jl,
+  author       = {Sanders, David P. and Benet, Luis},
+  title        = {ValidatedNumerics.jl: Rigorous floating-point calculations
+      with interval arithmetic in Julia},
+  howpublished = {\url{https://github.com/JuliaIntervals/ValidatedNumerics.jl}}
+}
+
+@misc{ArbFloats.jl,
+  author       = {Sarnoff, Jeffrey},
+  title        = {ArbFloats.jl},
+  howpublished = {\url{https://github.com/JuliaArbTypes/ArbFloats.jl}}
+}
+@misc{TaylorIntegration.jl,
+  author       = {P\'erez-Hern\'andez, Jorge A. and Benet, Luis},
+  title        = {TaylorIntegration.jl: precise ODE integration using Taylor's
+      method in Julia },
+  howpublished = {\url{https://github.com/PerezHz/TaylorIntegration.jl}}
+}
+
+@misc{TaylorModels.jl,
+  author       = {Benet, Luis and Sanders, David P.},
+  title        = {TaylorModels.jl: Rigorous function approximation with Taylor
+    models in Julia},
+  howpublished = {\url{https://github.com/PerezHz/TaylorIntegration.jl}}
+}
