### 2018-09-22 #968 @lbenet
Thanks @jedbrown! I just merged https://github.com/JuliaDiff/TaylorSeries.jl/pull/181, so the paper is now in master. Sorry for the confusion.

Regarding the referees, @ChrisRackauckas, @vchuravy and/or @datseris are certainly good options.
