### 2018-09-24 71d18a3
Add outer constructor
