### 2018-09-24 8a9348f
Fix new constructor
