### 2018-09-24 726d4cf
Add non-parametric Taylor1 outer constructor
