### 2018-09-24 6c499fc
Add `inv` test
