+++ test/onevariable.jl
@@ -480,6 +480,13 @@ eeuler = Base.MathConstants.e
     for i in 0:2
         @test a[i]*(180/pi) == b[i]
     end
+
+    a = rand(3, 3)
+    b = Taylor1.(a, 5)
+    binv = inv(b)
+    I_t1_5 = Taylor1.(Matrix{Float64}(I, size(b)), 5) # 5x5 Taylor1{Float64} matrix, order 5
+    @test norm(b*binv - I_t1_5, Inf) ≤ 1e-14
+    @test norm(binv*b - I_t1_5, Inf) ≤ 1e-14
 end
 
 @testset "Matrix multiplication for Taylor1" begin
