### 2018-09-24 17d1a23
Add constructor tests
