### 2018-09-25 bdbcdbb
Address PR review - Remove constructor method - Fix tests
