### 2018-09-25 f8e6617
Add Taylor1 constructors; fix #183 (#184)
* Add outer constructor

* Fix new constructor

* Add non-parametric Taylor1 outer constructor

* Fix

* Add `inv` test

* Add constructor tests

* Fix comment
[skip ci]

* Address PR review
- Remove constructor method
- Fix tests
