+++ test/onevariable.jl
@@ -480,6 +480,20 @@ eeuler = Base.MathConstants.e
     for i in 0:2
         @test a[i]*(180/pi) == b[i]
     end
+
+    # Test additional Taylor1 constructors
+    @test Taylor1{Float64}(true) == Taylor1([1.0])
+    @test Taylor1{Float64}(false) == Taylor1([0.0])
+    @test Taylor1{Int}(true) == Taylor1([1])
+    @test Taylor1{Int}(false) == Taylor1([0])
+
+    # Test use of `inv` for computation of matrix inverse of `Matrix{Taylor1{Float64}}`
+    a = rand(3, 3)
+    b = Taylor1.(a, 5)
+    binv = inv(b)
+    I_t1_5 = Taylor1.(Matrix{Float64}(I, size(b)), 5) # 5x5 Taylor1{Float64} identity matrix, order 5
+    @test norm(b*binv - I_t1_5, Inf) ≤ 1e-12
+    @test norm(binv*b - I_t1_5, Inf) ≤ 1e-12
 end
 
 @testset "Matrix multiplication for Taylor1" begin
