### 2018-10-02 #968 @jedbrown
@lbenet It looks like you missed `.jl` in some bib keys. I'll expand my reviewer search.
