### 2018-10-02 9957618
Fix references in paper.md
