+++ paper/paper.md
@@ -25,7 +25,7 @@ bibliography: paper.bib
 [`TaylorSeries.jl`](https://github.com/JuliaDiff/TaylorSeries.jl)
 [@TaylorSeries] provides a framework to use and manipulate
 Taylor polynomials in one and more variables
-in the [Julia programming language](https://julialang.org) [@Julia]. It allows to compute
+in the [Julia programming language](https://julialang.org) [@julia]. It allows to compute
 elementary functions, e.g. `exp`, of
 polynomials (`Taylor1`- or `TaylorN`-type objects), where
 techniques of automatic differentiation are used
