### 2018-10-02 #968 @lbenet
Thanks again; I just pushed the corrections to branch lbenet/corrections_paper; I also created a new PR, though I do not know if it should be merged now, or wait for the review.
