### 2018-10-17 5b1c2ad
Cap Documenter.jl to 0.19 on Travis
Documenter 0.20 will introduce breaking changes that will invalidate
existing make.jl setups. This commit makes sure that automatic Travis
builds will not use 0.20 automatically, in order to avoid sudden
documentation deployment failures once Documenter 0.20 gets tagged.

This commit has been generated by a script.
