### 2018-10-22 c272d07
Corrections paper (keys in bibliography) (#185)
* Fix references in paper.md

* One more fix in the references
