### 2018-10-22 #1043 @whedon
[REVIEW]: TaylorSeries.jl: Taylor expansions in one and many variables in Julia
**Submitting author:** @lbenet (<a href="http://orcid.org/0000-0002-8470-9054">Luis Benet</a>)
**Repository:** <a href="https://github.com/JuliaDiff/TaylorSeries.jl" target ="_blank">https://github.com/JuliaDiff/TaylorSeries.jl</a>
**Version:** v0.9.2
