### 2018-10-22 #968 @lbenet
Thanks @jedbrown!

A small note: today the last version of the paper (with the corrected links of the references, https://github.com/JuliaDiff/TaylorSeries.jl/pull/185) was merged into master.
