### 2018-11-24 #1043 @tobydriscoll
Thirty-hour days? 😁

I should be on track to do it in the upcoming week.

---
Tobin Driscoll
Professor of Mathematical Sciences
Director of Undergraduate Studies
Director, Center for Applications of Mathematics in Medicine
University of Delaware, Newark, DE 19716
Twitter: @tobydriscoll
Site: http://www.tobydriscoll.net
