### 2018-11-30 #1043 @tobydriscoll
It's excellent software, well documented and with all the required elements in place. I have just minor (very picky) observations:

1. There should be some pointers to other automatic differentiation software for Julia. It's not hard to find, but links are appreciated.
1. The display of exponents is partially obscured for me in Juno. It's find at a command shell.
1. Maybe reconsider exporting the names "gradient" and "hessian", as those are not uncommon from other packages too. But I recognize that namespace is becoming a thorny Julia problem.
1. The function names "integrate" and "derivative" are verb and noun. I would have preferred they both be one or the other. (I did say I was being picky here.)
