### 2018-12-07 fcde7d0
Do not export gradient, jacobian and hessian
