+++ src/TaylorSeries.jl
@@ -22,7 +22,7 @@ using SparseArrays: SparseMatrixCSC
 using Markdown
 import LinearAlgebra: norm, mul!
 if VERSION ≥ v"1.0.0"
-    export gradient
+    # export gradient
 else
     import LinearAlgebra: gradient
 end
@@ -49,7 +49,8 @@ export getcoeff, derivative, integrate,
     get_order, get_numvars,
     set_variables, get_variables,
     get_variable_names, get_variable_symbols,
-    ∇, jacobian, jacobian!, hessian, hessian!,
+    # jacobian, hessian,
+    ∇, jacobian!, hessian!,
     taylor_expand, update!, constant_term
 
 include("parameters.jl")
