### 2018-12-07 155ad6f
Add other automatic diff packages in docs
