### 2018-12-12 9fef9d8
Add differentiate (synonym of derivative)
