+++ src/TaylorSeries.jl
@@ -43,7 +43,7 @@ import Base: zero, one, zeros, ones, isinf, isnan, iszero,
 
 export Taylor1, TaylorN, HomogeneousPolynomial, AbstractSeries
 
-export getcoeff, derivative, integrate,
+export getcoeff, derivative, integrate, differentiate,
     evaluate, evaluate!, inverse,
     show_params_TaylorN, show_monomials, displayBigO, use_show_default,
     get_order, get_numvars,
