+++ docs/src/userguide.md
@@ -138,7 +138,8 @@ rationalize(expon[3])
 ```
 
 Differentiating and integrating is straightforward for polynomial expansions in
-one variable, using [`derivative`](@ref) and [`integrate`](@ref). These
+one variable, using [`derivative`](@ref) and [`integrate`](@ref). (The
+function [`differentiate`](@ref) is an exact synonym of `derivative`.) These
 functions return the corresponding [`Taylor1`](@ref) expansions.
 The last coefficient of a derivative is set to zero to keep the
 same order as the original polynomial; for the integral, an
@@ -153,7 +154,7 @@ integrate(exp(t))
 integrate( exp(t), 1.0)
 integrate( derivative( exp(-t)), 1.0 ) == exp(-t)
 derivative(1, exp(shift_taylor(1.0))) == exp(1.0)
-derivative(5, exp(shift_taylor(1.0))) == exp(1.0) # 5-th derivative of `exp(1+t)`
+differentiate(5, exp(shift_taylor(1.0))) == exp(1.0) # 5-th derivative of `exp(1+t)`
 ```
 
 To evaluate a Taylor series at a given point, Horner's rule is used via the
