### 2019-01-07 04f8b1e
Unexport jacobian! and hessian!
