+++ test/manyvariables.jl
@@ -471,9 +471,9 @@ eeuler = Base.MathConstants.e
     @test ∇(f2) == [2*xT - 4*xT^3, TaylorN(1,0)]
     @test TaylorSeries.jacobian([f1,f2], [2,1]) == TaylorSeries.jacobian( [g1(xT+2,yT+1), g2(xT+2,yT+1)] )
     jac = Array{Int64}(undef, 2, 2)
-    jacobian!(jac, [g1(xT+2,yT+1), g2(xT+2,yT+1)])
+    TaylorSeries.jacobian!(jac, [g1(xT+2,yT+1), g2(xT+2,yT+1)])
     @test jac == TaylorSeries.jacobian( [g1(xT+2,yT+1), g2(xT+2,yT+1)] )
-    jacobian!(jac, [f1,f2], [2,1])
+    TaylorSeries.jacobian!(jac, [f1,f2], [2,1])
     @test jac == TaylorSeries.jacobian([f1,f2], [2,1])
     @test TaylorSeries.hessian( f1*f2 ) ==
         [derivative((2,0), f1*f2) derivative((1,1), (f1*f2));
@@ -486,16 +486,16 @@ eeuler = Base.MathConstants.e
     @test TaylorSeries.hessian(f1-f2-2*f1*f2) == (TaylorSeries.hessian(f1-f2-2*f1*f2))'
     @test TaylorSeries.hessian(f1-f2,[1,-1]) == TaylorSeries.hessian(g1(xT+1,yT-1)-g2(xT+1,yT-1))
     hes = Array{Int64}(undef, 2, 2)
-    hessian!(hes, f1*f2)
+    TaylorSeries.hessian!(hes, f1*f2)
     @test hes == TaylorSeries.hessian(f1*f2)
     @test [xT yT]*hes*[xT, yT] == [ 2*TaylorN((f1*f2)[2]) ]
-    hessian!(hes, f1^2)
+    TaylorSeries.hessian!(hes, f1^2)
     @test hes/2 == [ [49,0] [0,12] ]
-    hessian!(hes, f1-f2-2*f1*f2)
+    TaylorSeries.hessian!(hes, f1-f2-2*f1*f2)
     @test hes == hes'
     hes1 = Array{Int64}(undef, 2, 2)
-    hessian!(hes1, f1-f2,[1,-1])
-    hessian!(hes, g1(xT+1,yT-1)-g2(xT+1,yT-1))
+    TaylorSeries.hessian!(hes1, f1-f2,[1,-1])
+    TaylorSeries.hessian!(hes, g1(xT+1,yT-1)-g2(xT+1,yT-1))
     @test hes1 == hes
 
     use_show_default(true)
