### 2019-01-09 0495404
Address observations to paper (#188)
* Do not export gradient, jacobian and hessian

* Add other automatic diff packages in docs

* Add differentiate (synonym of derivative)

* Unexport jacobian! and hessian!

* Include `import Pkg` in coverage evaluation (travis)
