+++ test/mixtures.jl
@@ -267,5 +267,5 @@ using LinearAlgebra, SparseArrays
     Pv = [rndTN(get_order(), 3), rndTN(get_order(), 3)]
     Qv = convert.(Taylor1{TaylorN{Float64}}, Pv)
 
-    @test jacobian(Pv) == jacobian(Qv)
+    @test TaylorSeries.jacobian(Pv) == TaylorSeries.jacobian(Qv)
 end
