+++ .travis.yml
@@ -23,6 +23,6 @@ script:
 after_success:
   - julia -e 'import Pkg; ps=Pkg.PackageSpec(name="Documenter", version="0.19"); Pkg.add(ps); Pkg.pin(ps)'
   - julia -e 'import TaylorSeries; include(joinpath(dirname(pathof(TaylorSeries)), "../docs", "make.jl"))'
-  - julia -e 'Pkg.add("Coverage"); using Coverage; Coveralls.submit(Coveralls.process_folder())'
+  - julia -e 'import Pkg; Pkg.add("Coverage"); using Coverage; Coveralls.submit(Coveralls.process_folder())'
 
 sudo: false
