+++ latest/search_index.js
@@ -45,7 +45,7 @@ var documenterSearchIndex = {"docs": [
     "page": "Home",
     "title": "Related packages",
     "category": "section",
-    "text": "Polynomials.jl: Polynomial manipulations\nPowerSeries.jl: Truncated power series for Julia\nMultivariatePolynomials.jl: Multivariate polynomials interface\nAbstractAlgebra.jl: Generic abstract algebra functionality in pure Julia"
+    "text": "Polynomials.jl: Polynomial manipulations\nPowerSeries.jl: Truncated power series for Julia\nMultivariatePolynomials.jl: Multivariate polynomials interface\nAbstractAlgebra.jl: Generic abstract algebra functionality in pure Julia\nForwardDiff.jl: Forward Mode Automatic Differentiation for Julia\nReverseDiff.jl: Reverse Mode Automatic Differentiation for Julia\nHyperDualNumbers.jl: Julia implementation of HyperDualNumbers"
 },
 
 {
@@ -125,7 +125,7 @@ var documenterSearchIndex = {"docs": [
     "page": "User guide",
     "title": "One independent variable",
     "category": "section",
-    "text": "Taylor expansions in one variable are represented by the Taylor1 type, which consists of a vector of coefficients (fieldname coeffs) and the maximum order considered for the expansion (fieldname order). The coefficients are arranged in ascending order with respect to the degree of the monomial, so that coeffs[1] is the constant term, coeffs[2] gives the first order term (t^1), etc. Yet, it is possible to have the natural ordering with respect to the degree; see below. This is a dense representation of the polynomial. The order of the polynomial can be omitted in the constructor, which is then fixed by the length of the vector of coefficients. If the length of the vector does not correspond with the order, order is used, which effectively truncates polynomial to degree order.Taylor1([1, 2, 3],4) # Polynomial of order 4 with coefficients 1, 2, 3\nTaylor1([0.0, 1im]) # Also works with complex numbers\nTaylor1(ones(8), 2) # Polynomial truncated to order 2\nshift_taylor(a) = a + Taylor1(typeof(a),5)  ## a + taylor-polynomial of order 5\nt = shift_taylor(0.0) # Independent variable `t`Note that the information about the maximum order considered is displayed using a big-𝒪 notation. In some cases, it is desirable to not display the big-𝒪 notation. The function displayBigO allows to control whether it is displayed or not.displayBigO(false) # turn-off displaying big O notation\nt\ndisplayBigO(true) # turn it on\ntSimilarly, it is possible to control if the format of the displayed series through the function use_show_default; use_show_default(true) uses the Base.show_default, while use_show_default(false) uses the custom display form (default).use_show_default(true) # use Base.show method\nt\nuse_show_default(false) # use custom `show`\ntThe definition of shift_taylor(a) uses the method Taylor1([::Type{Float64}], [order::Int64=1]), which is a shortcut to define the independent variable of a Taylor expansion, of given type and order (defaults are Float64 and order=1). This is one of the easiest ways to work with the package.The usual arithmetic operators (+, -, *, /, ^, ==) have been extended to work with the Taylor1 type, including promotions that involve Numbers. The operations return a valid Taylor expansion of maximum order. This is apparent in the last example below, where the answer is beyond the order of the expansion.t*(3t+2.5)\n1/(1-t)\nt*(t^2-4)/(t+2)\ntI = im*t\n(1-t)^3.2\n(1+t)^t\nt^6  # t is of order 5If no valid Taylor expansion can be computed, an error is thrown, for instance when a derivative is not defined (or simply diverges):1/t\nt^3.2\nabs
[…]
