### 2019-01-09 #1043 @lbenet
@tobydriscoll We have just merged https://github.com/JuliaDiff/TaylorSeries.jl/pull/188, which addresses your comments.

The concrete changes are the following:
- We added links to other Julia packages related to automatic differentiation in the documentation.
- We are not exporting any more `gradient`, `jacobian`, `hessian`, `jacobian!` and `hessian!` as suggested. A new version will be released that includes these changes.
- We are keeping `derivative` but added `differentiate`.
- The documentation was updated.

Regarding to point 2 in your review, we couldn't reproduce the problem. It may be related to the fonts you have, or to the specific platform; if you can provide us with more details, we could address the problem.

Thanks a lot for the review. If you have further suggestions, please let us know them.

cc @jedbrown
