### 2019-01-21 6c6d627
Int64 -> Int
This follows from a suggestion by @mforets TaylorModels.jl#16
