+++ src/other_functions.jl
@@ -210,19 +210,19 @@ a `TaylorN` expansion will be computed. If the dimension of x0 (`length(x0)`)
 is different from the variables set for `TaylorN` (`get_numvars()`), an
 `AssertionError` will be thrown.
 """
-function taylor_expand(f::Function; order::Int64=15)
+function taylor_expand(f::Function; order::Int=15)
    a = Taylor1(order)
    return f(a)
 end
 
-function taylor_expand(f::Function, x0::T; order::Int64=15) where {T<:Number}
+function taylor_expand(f::Function, x0::T; order::Int=15) where {T<:Number}
    a = Taylor1([x0, one(T)], order)
    return f(a)
 end
 
 #taylor_expand function for TaylorN
 function taylor_expand(f::Function, x0::Vector{T};
-        order::Int64=get_order()) where {T<:Number}
+        order::Int=get_order()) where {T<:Number}
 
     ll = length(x0)
     @assert ll == get_numvars() && order <= get_order()
@@ -235,7 +235,7 @@ function taylor_expand(f::Function, x0::Vector{T};
     return f( X )
 end
 
-function taylor_expand(f::Function, x0...; order::Int64=get_order())
+function taylor_expand(f::Function, x0...; order::Int=get_order())
     x0 = promote(x0...)
     T = eltype(x0[1])
     ll = length(x0)
