### 2019-01-21 506facd
Int64 -> Int (#189)
This follows from a suggestion by @mforets TaylorModels.jl#16
