+++ versions.js
@@ -5,6 +5,7 @@ var DOC_VERSIONS = [
   "release-0.7",
   "release-0.6",
   "release-0.5",
+  "v0.8.1",
   "v0.8.0",
   "v0.7.4",
   "v0.7.3",
