### 2019-01-24 a6463c2
Test also Julia 1.1
