+++ .travis.yml
@@ -7,6 +7,7 @@ os:
 julia:
     - 0.7
     - 1.0
+    - 1.1
     - nightly
 
 notifications:
@@ -14,15 +15,23 @@ notifications:
 
 matrix:
   allow_failures:
-  - julia: nightly
+   - julia: nightly
 
-script:
- - if [[ -a .git/shallow ]]; then git fetch --unshallow; fi
- - julia --project --check-bounds=yes -e 'using Pkg; Pkg.clone(pwd()); Pkg.add("TaylorSeries"); Pkg.test("TaylorSeries"; coverage=true)'
+env:
+  global:
+    - DOCUMENTER_DEBUG=true
+
+jobs:
+  include:
+    - stage: "Documentation"
+      julia: 1.0
+      os: linux
+      script:
+        - julia --project=docs/ -e 'using Pkg; Pkg.develop(PackageSpec(path=pwd())); Pkg.instantiate()'
+        - julia --project=docs/ docs/make.jl
+      after_success: skip
 
 after_success:
-  - julia -e 'import Pkg; ps=Pkg.PackageSpec(name="Documenter", version="0.19"); Pkg.add(ps); Pkg.pin(ps)'
-  - julia -e 'import TaylorSeries; include(joinpath(dirname(pathof(TaylorSeries)), "../docs", "make.jl"))'
   - julia -e 'import Pkg; Pkg.add("Coverage"); using Coverage; Coveralls.submit(Coveralls.process_folder())'
 
 sudo: false
