### 2019-01-24 c1253dc
Add docs/Project.toml and remove docs/REQUIRE
