### 2019-01-24 19a0939
Build the docs with Julia 1.1, and solve warnings
