### 2019-01-24 #1043 @lbenet
@jedbrown I'd like to mention that we released today v0.8.1 of our package. I hope this is ok with the current review process.
