### 2019-01-31 ba50e6d
Rewrite integrate for Taylor1, avoiding  promote_type (#191)
