### 2019-02-01 6f80e93
Test and build docs using Julia v1.1 (#190)
* Test also Julia 1.1

* Add docs/Project.toml and remove docs/REQUIRE

* Build the docs with Julia 1.1, and solve warnings
