+++ dev/search_index.js
@@ -1,7 +1,7 @@
 var documenterSearchIndex = {"docs": [
 
 {
-    "location": "index.html#",
+    "location": "#",
     "page": "Home",
     "title": "Home",
     "category": "page",
@@ -9,7 +9,7 @@ var documenterSearchIndex = {"docs": [
 },
 
 {
-    "location": "index.html#TaylorSeries.jl-1",
+    "location": "#TaylorSeries.jl-1",
     "page": "Home",
     "title": "TaylorSeries.jl",
     "category": "section",
@@ -17,7 +17,7 @@ var documenterSearchIndex = {"docs": [
 },
 
 {
-    "location": "index.html#Authors-1",
+    "location": "#Authors-1",
     "page": "Home",
     "title": "Authors",
     "category": "section",
@@ -25,7 +25,7 @@ var documenterSearchIndex = {"docs": [
 },
 
 {
-    "location": "index.html#License-1",
+    "location": "#License-1",
     "page": "Home",
     "title": "License",
     "category": "section",
@@ -33,7 +33,7 @@ var documenterSearchIndex = {"docs": [
 },
 
 {
-    "location": "index.html#Installation-1",
+    "location": "#Installation-1",
     "page": "Home",
     "title": "Installation",
     "category": "section",
@@ -41,7 +41,7 @@ var documenterSearchIndex = {"docs": [
 },
 
 {
-    "location": "index.html#Related-packages-1",
+    "location": "#Related-packages-1",
     "page": "Home",
     "title": "Related packages",
     "category": "section",
@@ -49,7 +49,7 @@ var documenterSearchIndex = {"docs": [
 },
 
 {
-    "location": "index.html#Acknowledgments-1",
+    "location": "#Acknowledgments-1",
     "page": "Home",
     "title": "Acknowledgments",
     "category": "section",
@@ -57,7 +57,7 @@ var documenterSearchIndex = {"docs": [
 },
 
 {
-    "location": "background.html#",
+    "location": "background/#",
     "page": "Background",
     "title": "Background",
     "category": "page",
@@ -65,7 +65,7 @@ var documenterSearchIndex = {"docs": [
 },
 
 {
-    "location": "background.html#Background-1",
+    "location": "background/#Background-1",
     "page": "Background",
     "title": "Background",
     "category": "section",
@@ -73,7 +73,7 @@ var documenterSearchIndex = {"docs": [
 },
 
 {
-    "location": "background.html#Introduction-1",
+    "location": "background/#Introduction-1",
     "page": "Background",
     "title": "Introduction",
     "category": "section",
@@ -81,7 +81,7 @@ var documenterSearchIndex = {"docs": [
 },
 
 {
-    "location": "background.html#Arithmetic-operations-1",
+    "location": "background/#Arithmetic-operations-1",
     "page": "Background",
     "title": "Arithmetic operations",
     "category": "section",
@@ -89,7 +89,7 @@ var documenterSearchIndex = {"docs": [
 },
 
 {
-    "location": "background.html#Elementary-functions-of-polynomials-1",
+    "location": "background/#Elementary-functions-of-polynomials-1",
     "page": "Background",
     "title": "Elementary functions of polynomials",
     "category": "section",
@@ -97,7 +97,7 @@ var documenterSearchIndex = {"docs": [
 },
 
 {
-    "location": "background.html#refs-1",
+    "location": "background/#refs-1",
     "page": "Background",
     "title": "References",
[…]
