+++ stable/search_index.js
@@ -1,763 +1,0 @@
-var documenterSearchIndex = {"docs": [
-
-{
-    "location": "index.html#",
-    "page": "Home",
-    "title": "Home",
-    "category": "page",
-    "text": ""
-},
-
-{
-    "location": "index.html#TaylorSeries.jl-1",
-    "page": "Home",
-    "title": "TaylorSeries.jl",
-    "category": "section",
-    "text": "A Julia package for Taylor expansions in one or more independent variables."
-},
-
-{
-    "location": "index.html#Authors-1",
-    "page": "Home",
-    "title": "Authors",
-    "category": "section",
-    "text": "Luis Benet, Instituto de Ciencias   Físicas, Universidad Nacional Autónoma de México (UNAM).\nDavid P. Sanders, Facultad   de Ciencias, Universidad Nacional Autónoma de México (UNAM)."
-},
-
-{
-    "location": "index.html#License-1",
-    "page": "Home",
-    "title": "License",
-    "category": "section",
-    "text": "TaylorSeries is licensed under the MIT \"Expat\" license; see LICENSE for the full license text."
-},
-
-{
-    "location": "index.html#Installation-1",
-    "page": "Home",
-    "title": "Installation",
-    "category": "section",
-    "text": "TaylorSeries.jl is a registered package, and is simply installed by runningpkg> add(\"TaylorSeries\")"
-},
-
-{
-    "location": "index.html#Related-packages-1",
-    "page": "Home",
-    "title": "Related packages",
-    "category": "section",
-    "text": "Polynomials.jl: Polynomial manipulations\nPowerSeries.jl: Truncated power series for Julia\nMultivariatePolynomials.jl: Multivariate polynomials interface\nAbstractAlgebra.jl: Generic abstract algebra functionality in pure Julia\nForwardDiff.jl: Forward Mode Automatic Differentiation for Julia\nReverseDiff.jl: Reverse Mode Automatic Differentiation for Julia\nHyperDualNumbers.jl: Julia implementation of HyperDualNumbers"
-},
-
-{
-    "location": "index.html#Acknowledgments-1",
-    "page": "Home",
-    "title": "Acknowledgments",
-    "category": "section",
-    "text": "This project began (using Python) during a Masters\' course in the postgraduate programs in Physics and in Mathematics at UNAM, during the second half of 2013. We thank the participants of the course for putting up with the half-baked material and contributing energy and ideas.We acknowledge financial support from DGAPA-UNAM PAPIME grants PE-105911 and PE-107114, and PAPIIT grants IG-101113 and IG-100616. LB acknowledges support through a Cátedra Marcos Moshinsky (2013)."
-},
-
-{
-    "location": "background.html#",
-    "page": "Background",
-    "title": "Background",
-    "category": "page",
-    "text": ""
-},
-
-{
-    "location": "background.html#Background-1",
-    "page": "Background",
-    "title": "Background",
-    "category": "section",
-    "text": ""
-},
-
-{
-    "location": "background.html#Introduction-1",
-    "page": "Background",
-    "title": "Introduction",
-    "category": "section",
[…]
