+++ versions.js
@@ -1,21 +1,11 @@
 var DOC_VERSIONS = [
   "stable",
-  "latest",
-  "release-0.8",
-  "release-0.7",
-  "release-0.6",
-  "release-0.5",
-  "v0.8.1",
-  "v0.8.0",
-  "v0.7.4",
-  "v0.7.3",
-  "v0.7.2",
-  "v0.7.1",
-  "v0.6.0",
-  "v0.5.1",
-  "v0.5.0",
-  "v0.3.0",
-  "v0.2.1",
-  "v0.2.0",
-  "v0.1.3",
+  "v0.8",
+  "v0.7",
+  "v0.6",
+  "v0.5",
+  "v0.3",
+  "v0.2",
+  "v0.1",
+  "dev",
 ];
