### 2019-02-01 12bad60
paper: bib updates and citing style
