+++ paper/paper.bib
@@ -1,6 +1,6 @@
 @misc{TaylorSeries,
   author       = {Benet, Luis and Sanders, David P.},
-  title        = {A julia package for Taylor expansions in one or
+  title        = {A {Julia} package for {Taylor} expansions in one or
     more independent variables},
   howpublished = {\url{https://github.com/JuliaDiff/TaylorSeries.jl}}
 }
@@ -32,9 +32,10 @@
     rigorous results to effective computations},
   publisher = {Springer International Publishing},
   address   = {Switzerland},
-  series    = {Applied mathematical sciences},
+  series    = {Applied Mathematical Sciences},
   year      = {2016},
   url       = {https://cds.cern.ch/record/2151752},
+  doi       = {10.1007/978-3-319-29662-3},
 }
 
 @article{MPFR,
@@ -59,7 +60,7 @@
 @misc{ValidatedNumerics.jl,
   author       = {Sanders, David P. and Benet, Luis},
   title        = {ValidatedNumerics.jl: Rigorous floating-point calculations
-      with interval arithmetic in Julia},
+      with interval arithmetic in {Julia}},
   howpublished = {\url{https://github.com/JuliaIntervals/ValidatedNumerics.jl}}
 }
 
@@ -70,14 +71,14 @@
 }
 @misc{TaylorIntegration.jl,
   author       = {P\'erez-Hern\'andez, Jorge A. and Benet, Luis},
-  title        = {TaylorIntegration.jl: precise ODE integration using Taylor's
-      method in Julia },
+  title        = {TaylorIntegration.jl: precise {ODE} integration using {Taylor}'s
+      method in {Julia}},
   howpublished = {\url{https://github.com/PerezHz/TaylorIntegration.jl}}
 }
 
 @misc{TaylorModels.jl,
   author       = {Benet, Luis and Sanders, David P.},
-  title        = {TaylorModels.jl: Rigorous function approximation with Taylor
-    models in Julia},
-  howpublished = {\url{https://github.com/PerezHz/TaylorIntegration.jl}}
+  title        = {TaylorModels.jl: Rigorous function approximation with {Taylor}
+    models in {Julia}},
+  howpublished = {\url{https://github.com/JuliaIntervals/TaylorModels.jl}}
 }
