### 2019-02-01 8cf19be
paper: bib updates and citing style
