### 2019-02-01 3a0da1b
paper: avoid self-cite; this info will be automatic
