### 2019-02-01 9fbc984
paper.bib: add year and adjust URL formatting
