+++ paper/paper.bib
@@ -61,24 +61,28 @@
   author       = {Sanders, David P. and Benet, Luis},
   title        = {ValidatedNumerics.jl: Rigorous floating-point calculations
       with interval arithmetic in {Julia}},
-  howpublished = {\url{https://github.com/JuliaIntervals/ValidatedNumerics.jl}}
+  url = {https://github.com/JuliaIntervals/ValidatedNumerics.jl},
+  year = {2019},
 }
 
 @misc{ArbFloats.jl,
   author       = {Sarnoff, Jeffrey},
   title        = {ArbFloats.jl},
-  howpublished = {\url{https://github.com/JuliaArbTypes/ArbFloats.jl}}
+  url = {https://github.com/JuliaArbTypes/ArbFloats.jl},
+  year = {2019},
 }
 @misc{TaylorIntegration.jl,
   author       = {P\'erez-Hern\'andez, Jorge A. and Benet, Luis},
   title        = {TaylorIntegration.jl: precise {ODE} integration using {Taylor}'s
       method in {Julia}},
-  howpublished = {\url{https://github.com/PerezHz/TaylorIntegration.jl}}
+  url = {https://github.com/PerezHz/TaylorIntegration.jl},
+  year = {2019},
 }
 
 @misc{TaylorModels.jl,
   author       = {Benet, Luis and Sanders, David P.},
   title        = {TaylorModels.jl: Rigorous function approximation with {Taylor}
     models in {Julia}},
-  howpublished = {\url{https://github.com/JuliaIntervals/TaylorModels.jl}}
+  url = {https://github.com/JuliaIntervals/TaylorModels.jl},
+  year = {2019},
 }
