### 2019-02-01 #1043 @jedbrown
Thanks @sriharikrishna!

@lbenet Please see my PR referenced above. Once merged, I'll re-check the paper and if all looks good, we'll be ready to archive. Thanks.
