### 2019-02-02 #1043 @lbenet
Thanks a lot @sriharikrishna.

@jedbrown There is a small typo on a keyword in https://github.com/JuliaDiff/TaylorSeries.jl/pull/192. Aside from that, LGTM.
