### 2019-02-04 #1043 @lbenet
https://github.com/JuliaDiff/TaylorSeries.jl/pull/192 is merged. Thanks a lot!
