### 2019-02-04 #1043 @lbenet
@jedbrown What is the usual way of proceeding now: tagging a new minor version (0.9.0), or a patch version (0.8.2)?
