### 2019-02-04 64bf5ac
use promotion in Taylor1's evaluate
