### 2019-02-05 #1043 @lbenet
@jedbrown Archiving done! The DOI is [10.5281/zenodo.2557004](https://zenodo.org/record/2557004).
