### 2019-02-05 2bc98ce
Add DOI badge (Zenodo)
