### 2019-02-05 #1043 @jedbrown
Looks good. In the future, I would recommend using [annotated tags](https://git-scm.com/book/en/v2/Git-Basics-Tagging). I think you could change it on GitHub and the Zenodo link would still work; we don't need a new DOI even if you choose to upgrade to an annotated tag.
