### 2019-02-05 #1043 @kyleniemeyer
Hi @jedbrown, I'm actually going to send this back to you (and the authors) for now. Right now the article itself is extremely brief and doesn't contain much substance; while we don't expect a full article here, it should be 250–1000 words (currently the main article is 147, which put together is a long paragraph). To me, these items need to be improved:
- A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience
- A clear statement of need that illustrates the purpose of the software

Further, I think the article should be improved with more description of how the software works, some examples, and use cases.
