### 2019-02-05 9f19aaf
add improved evaluation for intervals
