### 2019-02-06 e5c338b
Additions to the paper
