+++ paper/paper.md
@@ -22,16 +22,27 @@ bibliography: paper.bib
 
 # Summary
 
-`TaylorSeries.jl`
-provides a framework to use and manipulate
+The purpose of `TaylorSeries.jl` is to provide a framework to exploit
 Taylor polynomials in one and more variables
-in the [Julia programming language](https://julialang.org) [@julia]. It allows to compute
-elementary functions, e.g. `exp`, of
-polynomials (`Taylor1`- or `TaylorN`-type objects), where
-techniques of automatic differentiation are used
+in the [Julia programming language](https://julialang.org) [@julia]. In
+certain cases, it provides a very primitive CAS (computer algebra system),
+though it does the work numerically, and not on a symbolic level;
+some examples can be found in the
+[documentation](http://www.juliadiff.org/TaylorSeries.jl/stable/).
+The package allows to manipulate polynomials of an specified maximum
+degree, including power and composition, as well as series expansions
+of some elementary functions of polynomials, e.g. `exp`,
+where techniques of automatic differentiation are used
 [@Tucker:ValidatedNumerics; @HaroEtAl:ParameterizMeth]. Differentiation and
 integration are also implemented.
 
+Two basic immutable types are defined, `Taylor1` and `TaylorN`,
+which represent the series expansions in one of more variables,
+respectively. These structures are essentially vectors of coefficients
+ordered increasingly by its degree. In the case of `TaylorN`, the
+coefficients are `HomogeneousPolynomials`, which in turn are vectors
+of coefficients representing all monomials of given number of variables
+and order (total degree), in some lexicographical ordering.
 The package allows to work with different `Number` formats
 as coefficients of the series, including complex numbers,
 the arbitrary precision `BigFloat`s [@MPFR],
@@ -42,10 +53,11 @@ as well as `Taylor1` and `TaylorN` objects.
 [`TaylorIntegration.jl`](https://github.com/PerezHz/TaylorIntegration.jl)
 [@TaylorIntegration.jl], whose aim is to perform accurate integration
 of ODEs using the Taylor method, including jet transport techniques,
-and of
+where a small region around an initial condition is integrated.
+It is also an important component of
 [`TaylorModels.jl`](https://github.com/JuliaIntervals/TaylorModels.jl)
-[@TaylorModels.jl],
-which constructs rigorous polynomial approximations of functions.
+[@TaylorModels.jl], whose aim is to construct rigorous polynomial
+approximations of functions.
 
 ## Acknowledgements
 
