### 2019-02-12 9cd952d
use new intersect method and fix normalization
