### 2019-02-15 371b79e
Add a normalization wrapper (normalize_taylor) and use dispatch
