### 2019-02-15 223c2dc
Export normalize_taylor
