### 2019-02-15 5716c3c
Add specialized evaluate methods for Interval's and IntervalBox'es
