### 2019-02-15 8075443
Fix bug, clean-up, and minor additions, including tests
