### 2019-02-18 42d8912
Clean-up evaluate.jl
