### 2019-02-19 793ff60
Correct a mistake in _evaluate, and add two tests
