+++ test/intervals.jl
@@ -40,12 +40,15 @@ eeuler = Base.MathConstants.e
     @test (((x-y)^4)[4])(a × b) == -39 .. 81
 
     p4n = normalize_taylor(p4(x,y), a × b, true)
-    @test p4n((-1..1)×(-1..1)) ⊆ p4(a,b)
+    @test p4n((-1..1)×(-1..1)) ⊇ p4(a,b)
     p5n = normalize_taylor(p5(x,y), a × b, true)
-    @test p5n((-1..1)×(-1..1)) ⊆ p5(a,b)
+    @test p5n((-1..1)×(-1..1)) ⊇ p5(a,b)
 
     p4n = normalize_taylor(p4(x,y), a × b, false)
     @test p4n((0..1)×(0..1)) ⊆ p4(a,b)
     p5n = normalize_taylor(p5(x,y), a × b, false)
     @test p5n((0..1)×(0..1)) ⊆ p5(a,b)
+
+    @test evaluate(x*y^3, (-1..1)×(-1..1)) == (-1..1)
+    @test evaluate(x*y^2, (-1..1)×(-1..1)) == (-1..1)
 end
