### 2019-02-20 5ae33e0
Correct ranges in _evaluate
