### 2019-02-20 5fce36d
cleanup commented code of initial implementation
