### 2019-02-28 03e448e
Implement review suggestions of @dpsanders
