### 2019-02-28 09418a8
Correct docstrings
