### 2019-03-01 b5e518b
Fixes in _evaluate methods and tweaks in _normalize, and add tests
