### 2019-03-03 6171532
Fix a subtle error, improve evaluate (Taylor1s at -1..1) and add more tests
The case of Taylor1s at 0..1 is identic to Horner's rule
