+++ src/intervals.jl
@@ -2,55 +2,37 @@ using .IntervalArithmetic
 
 function evaluate(a::Taylor1, dx::Interval)
     dx == (-1..1) && return _evaluate(a, dx, Val(true))
-    dx == (0..1) && return _evaluate(a, dx, Val(false))
     # Usual Horner rule
     uno = one(dx)
     @inbounds begin
         suma = a[end]*uno
         for k in a.order-1:-1:0
-            suma = suma*dx + a[k]*uno
+            suma = suma*dx + a[k]
         end
     end
     return suma
 end
 
 function _evaluate(a::Taylor1, dx::Interval{T}, ::Val{true}) where {T}
+    uno = Interval{T}(1, 1)
+    dx2 = Interval{T}(0, 1)
     if iseven(a.order)
-        @inbounds begin
-            suma1 = a[end-1]
-            for k in a.order-3:-2:1
-                suma1 += a[k]
-            end
-            suma2 = a[end]
-            for k in a.order-2:-2:2
-                suma2 += a[k]
-            end
-        end
+        kend = a.order-2
+        @inbounds sum_even = a[end]*uno
+        @inbounds sum_odd = a[end-1]*Interval{T}(0, 0)
     else
-        @inbounds begin
-            suma1 = a[end]
-            for k in a.order-2:-2:1
-                suma1 += a[k]
-            end
-            suma2 = a[end-1]
-            for k in a.order-3:-2:2
-                suma2 += a[k]
-            end
-        end
+        kend = a.order-3
+        @inbounds sum_odd = a[end]*uno
+        @inbounds sum_even = a[end-1]*uno
     end
-    return a[0] + suma1*dx + suma2 * Interval{T}(0, 1)
+    @inbounds for k in kend:-2:0
+        sum_odd = sum_odd*dx2 + a[k+1]
+        sum_even = sum_even*dx2 + a[k]
+    end
+    return sum_even + sum_odd*dx
 end
 
-function _evaluate(a::Taylor1, dx::Interval, ::Val{false})
-    uno = one(dx)
-    @inbounds begin
-        suma = a[end]*uno
-        for k in a.order-1:-1:1
-            suma = suma*dx + a[k]*uno
-        end
-    end
-    return a[0] + suma
-end
+_evaluate(a::Taylor1, dx::Interval, ::Val{false}) = a(dx)
 
 
 function evaluate(a::TaylorN, dx::IntervalBox{N,T}) where {N,T}
