### 2019-03-03 1adfdda
Add two examples related to Hermite polynomials
