+++ paper/paper.md
@@ -26,26 +26,24 @@ The purpose of `TaylorSeries.jl` is to provide a framework to exploit
 Taylor polynomials in one and more variables
 in the [Julia programming language](https://julialang.org) [@julia]. In
 certain cases, it provides a very primitive CAS (computer algebra system),
-though it does the work numerically, and not on a symbolic level;
-some examples can be found in the
-[documentation](http://www.juliadiff.org/TaylorSeries.jl/stable/).
-The package allows to manipulate polynomials of an specified maximum
+which works numerically and not symbolically.
+The package allows to manipulate polynomials of a specified maximum
 degree, including power and composition, as well as series expansions
-of some elementary functions of polynomials, e.g. `exp`,
+of some elementary functions on polynomials, e.g. `exp`,
 where techniques of automatic differentiation are used
 [@Tucker:ValidatedNumerics; @HaroEtAl:ParameterizMeth]. Differentiation and
 integration are also implemented.
 
 Two basic immutable types are defined, `Taylor1` and `TaylorN`,
-which represent the series expansions in one of more variables,
+which represent the series expansions in one or several variables,
 respectively. These structures are essentially vectors of coefficients
 ordered increasingly by its degree. In the case of `TaylorN`, the
 coefficients are `HomogeneousPolynomials`, which in turn are vectors
 of coefficients representing all monomials of given number of variables
-and order (total degree), in some lexicographical ordering.
+and order (total degree), in some lexicographical order.
 The package allows to work with different `Number` formats
-as coefficients of the series, including complex numbers,
-the arbitrary precision `BigFloat`s [@MPFR],
+for the coefficients of the series, including complex numbers,
+arbitrary precision `BigFloat`s [@MPFR],
 `Interval`s [@ValidatedNumerics.jl], `ArbFloat`s [@ArbFloats.jl],
 as well as `Taylor1` and `TaylorN` objects.
 
@@ -59,6 +57,81 @@ It is also an important component of
 [@TaylorModels.jl], whose aim is to construct rigorous polynomial
 approximations of functions.
 
+# Examples
+
+We present three examples to show the use of `TaylorSeries.jl`. Other
+examples as well as a detailed user guide can be found in the
+[documentation](http://www.juliadiff.org/TaylorSeries.jl/stable/).
+
+As a first example we describe how to generate the [Hermite polynomials][@HermitePols]
+("physicists" version), up to a maximum order. We begin exploiting directly
+the recurrence relation satisfied by the polynomials.
+
+```julia
+julia> using TaylorSeries
+
+julia> displayBigO(false);
+
+julia> function generate_hermite_polynomials(::Type{T}, nmax::Int) where {T<:Integer}
+  Hn = Vector{Taylor1{T}}(undef, nmax+1)
+  Hn[1] = one(T)+Taylor1(T, 0) # order 0
+  Hn[2] = 2*Taylor1(T, 1)      # order 1
+  for n in 2:nmax
+      x = Taylor1(Int128, n) # Taylor variable of n-th order
+      # Recursion formula for degree n
+      Hn[n+1] = 2*x*Hn[n] - 2*(n-1)*Hn[n-1]
+  end
+  return Hn
+end;
+
+julia> generate_hermite_polynomials(n) = generate_hermite_polynomials(Int, n);
+
+julia> Hn = generate_hermite_polynomials(10);
+
+julia> function HermitePol(n::Int)
+    @assert 0 ≤ n ≤ length(Hn) "Not enough Hermite polynomials generated"
+    return Hn[n+1]
+end;
+
[…]
