+++ paper/paper.bib
@@ -2,7 +2,8 @@
   author       = {Benet, Luis and Sanders, David P.},
   title        = {A {Julia} package for {Taylor} expansions in one or
     more independent variables},
-  howpublished = {\url{https://github.com/JuliaDiff/TaylorSeries.jl}}
+  howpublished = {\url{https://github.com/JuliaDiff/TaylorSeries.jl}},
+  doi     = {10.5281/zenodo.2557003},
 }
 
 @article{julia,
@@ -71,12 +72,14 @@
   url = {https://github.com/JuliaArbTypes/ArbFloats.jl},
   year = {2019},
 }
+
 @misc{TaylorIntegration.jl,
   author       = {P\'erez-Hern\'andez, Jorge A. and Benet, Luis},
   title        = {TaylorIntegration.jl: precise {ODE} integration using {Taylor}'s
       method in {Julia}},
   url = {https://github.com/PerezHz/TaylorIntegration.jl},
   year = {2019},
+  doi = {10.5281/zenodo.2562364},
 }
 
 @misc{TaylorModels.jl,
