### 2019-03-03 88b35a5
Add the reference to Hermite polynomials (wikipedia)
