### 2019-03-04 0d2b67d
Fix a typo and minor changes
