### 2019-03-05 ee7cb52
Simplify Hermite recurrence code; minor changes to English
