### 2019-03-08 18ab85b
Extend broadcasted to all TaylorSeries types
