### 2019-03-08 3151b69
Changes to text of paper
