+++ paper/paper.bib
@@ -58,11 +58,10 @@
  keywords  = {IEEE 754 standard, Multiple-precision arithmetic, correct rounding, elementary function, floating-point arithmetic, portable software},
 }
 
-@misc{ValidatedNumerics.jl,
+@misc{IntervalArithmetic.jl,
   author       = {Sanders, David P. and Benet, Luis},
-  title        = {ValidatedNumerics.jl: Rigorous floating-point calculations
-      with interval arithmetic in {Julia}},
-  url = {https://github.com/JuliaIntervals/ValidatedNumerics.jl},
+  title        = {IntervalArithmetic.jl: Rigorous floating-point calculations using interval arithmetic in {Julia}},
+  url = {https://github.com/JuliaIntervals/IntervalArithmetic.jl},
   year = {2019},
 }
 
