### 2019-03-10 dd84ba8
Add improved evaluation for intervals (#195)
* add improved evaluation for intervals

* use new intersect method and fix normalization

* Add a normalization wrapper (normalize_taylor) and use dispatch

* Export normalize_taylor

* Add specialized evaluate methods for Interval's and IntervalBox'es

* Fix bug, clean-up, and minor additions, including tests

* Clean-up evaluate.jl

* Correct a mistake in _evaluate, and add two tests

* Correct ranges in _evaluate

* cleanup commented code of initial implementation

* Implement review suggestions of @dpsanders

* Correct docstrings

* minor tweak
[…]
