+++ test/intervals.jl
@@ -32,10 +32,49 @@ eeuler = Base.MathConstants.e
     for ind in eachindex(p5(x,-b))
         @test all((p5(x,-b)[ind]).coeffs .⊆ (((x-b)^5)[ind]).coeffs)
     end
-    @test evaluate(p4(x,-b), IntervalBox(a,b)) == p4(a, b)
-    @test (p5(x,-b))(IntervalBox(a,b)) == p5(a, b)
 
     # Tests `evaluate`
+    @test evaluate(p4(x,y), IntervalBox(a,-b)) == p4(a, -b)
+    @test (p5(x,y))(IntervalBox(a,b)) == p5(a, b)
     @test (a-b)^4 ⊆ ((x-y)^4)(a × b)
     @test (((x-y)^4)[4])(a × b) == -39 .. 81
+
+    p4n = normalize_taylor(p4(x,y), a × b, true)
+    @test (0..16) ⊆ p4n((-1..1)×(-1..1))
+    p5n = normalize_taylor(p5(x,y), a × b, true)
+    @test (-32 .. 32) ⊆ p5n((-1..1)×(-1..1))
+
+    p4n = normalize_taylor(p4(x,y), a × b, false)
+    @test (0..16) ⊆ p4n((0..1)×(0..1))
+    p5n = normalize_taylor(p5(x,y), a × b, false)
+    @test (0..32) ⊆ p5n((0..1)×(0..1))
+
+    @test evaluate(x*y^3, (-1..1)×(-1..1)) == (-1..1)
+    @test evaluate(x*y^2, (-1..1)×(-1..1)) == (-1..1)
+    @test evaluate(x^2*y^2, (-1..1)×(-1..1)) == (0..1)
+
+    ii = 0..6
+    t = Taylor1(4)
+    f(x) = 0.1 * x^3 - 0.5*x^2 + 1
+    ft = f(t)
+    f1 = normalize_taylor(ft, ii, true)
+    f2 = normalize_taylor(ft, ii, false)
+    @test Interval(-23/27, f(6)) ⊆ f(ii)
+    @test Interval(-23/27, f(6)) ⊆ ft(ii)
+    @test Interval(-23/27, f(6)) ⊆ f1(-1..1)
+    @test Interval(-23/27, f(6)) ⊆ f2(0..1)
+    @test f1(-1..1) ⊆ f(ii)
+    @test diam(f1(-1..1)) < diam(f2(0..1))
+
+    # An example from Makino's thesis
+    ii = 0..1
+    t = Taylor1(5)
+    g(x) = 1 - x^4 + x^5
+    gt = g(t)
+    g1 = normalize_taylor(gt, 0..1, true)
+    @test Interval(g(4/5),1) ⊆ g(ii)
+    @test Interval(g(4/5),1) ⊆ gt(ii)
+    @test Interval(g(4/5),1) ⊆ g1(-1..1)
+    @test g1(-1..1) ⊂ g(ii)
+    @test diam(g1(-1..1)) < diam(gt(ii))
 end
