+++ src/intervals.jl
@@ -1,0 +1,148 @@
+using .IntervalArithmetic
+
+function evaluate(a::Taylor1, dx::Interval)
+    dx == (-1..1) && return _evaluate(a, dx, Val(true))
+    # Usual Horner rule
+    uno = one(dx)
+    @inbounds begin
+        suma = a[end]*uno
+        for k in a.order-1:-1:0
+            suma = suma*dx + a[k]
+        end
+    end
+    return suma
+end
+
+function _evaluate(a::Taylor1, dx::Interval{T}, ::Val{true}) where {T}
+    uno = Interval{T}(1, 1)
+    dx2 = Interval{T}(0, 1)
+    if iseven(a.order)
+        kend = a.order-2
+        @inbounds sum_even = a[end]*uno
+        @inbounds sum_odd = a[end-1]*Interval{T}(0, 0)
+    else
+        kend = a.order-3
+        @inbounds sum_odd = a[end]*uno
+        @inbounds sum_even = a[end-1]*uno
+    end
+    @inbounds for k in kend:-2:0
+        sum_odd = sum_odd*dx2 + a[k+1]
+        sum_even = sum_even*dx2 + a[k]
+    end
+    return sum_even + sum_odd*dx
+end
+
+_evaluate(a::Taylor1, dx::Interval, ::Val{false}) = a(dx)
+
+
+function evaluate(a::TaylorN, dx::IntervalBox{N,T}) where {N,T}
+
+    @assert N == get_numvars()
+    a_length = length(a)
+    suma = zero(constant_term(a)) + Interval{T}(0, 0)
+    @inbounds for homPol in length(a):-1:1
+        suma += evaluate(a.coeffs[homPol], dx)
+    end
+
+    return suma
+end
+
+function evaluate(a::HomogeneousPolynomial, dx::IntervalBox{N,T}) where {T, N}
+
+    @assert N == get_numvars()
+    dx == IntervalBox(-1..1, Val(N)) && return _evaluate(a, dx, Val(true))
+    dx == IntervalBox( 0..1, Val(N)) && return _evaluate(a, dx, Val(false))
+
+    return evaluate(a, dx...)
+end
+
+function _evaluate(a::HomogeneousPolynomial, dx::IntervalBox{N,T}, ::Val{true} ) where {T, N}
+    a.order == 0 && return a[1] + Interval{T}(0, 0)
+
+    ct = coeff_table[a.order+1]
+    @inbounds suma = a[1]*Interval{T}(0,0)
+
+    Ieven = Interval{T}(0,1)
+    for (i,a_coeff) in enumerate(a.coeffs)
+        iszero(a_coeff) && continue
+        if isodd(sum(ct[i]))
+            suma += sum(a_coeff) * dx[1]
+            continue
+        end
+        @inbounds tmp = iseven(ct[i][1]) ? Ieven : dx[1]
+        for n in 2:N
+            @inbounds vv = iseven(ct[i][n]) ? Ieven : dx[1]
+            tmp *= vv
+        end
+        suma += a_coeff * tmp
+    end
+    return suma
+end
+
+function _evaluate(a::HomogeneousPolynomial, dx::IntervalBox{N,T}, ::Val{false} ) where {T, N}
+    a.order == 0 && return a[1] + Interval{T}(0, 0)
+
+    @inbounds suma = zero(a[1])*dx[1]
+    @inbounds for homPol in a.coeffs
+        suma += homPol*dx[1]
+    end
+    return suma
+end
+
+
+"""
+    normalize_taylor(a::Taylor1, I::Interval, symI::Bool=true)
+
+Normalizes `a::Taylor1` such that the interval `I` is mapped
[…]

[…]
