### 2019-03-12 27ec07e
Add Taylor integration example to paper
