+++ paper/paper.md
@@ -59,34 +59,34 @@ approximations of functions.
 
 # Examples
 
-We present three examples to show the use of `TaylorSeries.jl`. Other
+We present three examples to illustrate the use of `TaylorSeries.jl`. Other
 examples, as well as a detailed user guide, can be found in the
 [documentation](http://www.juliadiff.org/TaylorSeries.jl/stable).
 
-
+## Hermite polynomials
 As a first example we describe how to generate the [Hermite polynomials][@HermitePols_wikipedia]
 ("physicist's" version) up to a given maximum order. Firstly we directly exploit the recurrence relation satisfied by the polynomials.
 
 ```julia
 julia> using TaylorSeries
 
-julia> displayBigO(false);
+julia> displayBigO(false)
 
 julia> function hermite_polynomials(::Type{T}, nmax::Int) where {T <: Integer}
 
-         x = Taylor1(T, nmax)    # Taylor variable
-         H = fill(x, nmax + 1)   # vector of Taylor series to be overwritten
+           x = Taylor1(T, nmax)    # Taylor variable
+           H = fill(x, nmax + 1)   # vector of Taylor series to be overwritten
 
-         H[1] = 1   # order 0
-         H[2] = 2x  # order 1
+           H[1] = 1   # order 0
+           H[2] = 2x  # order 1
 
-         for n in 2:nmax
-             # recursion relation for order n:
-             H[n+1] = 2x * H[n] - 2(n-1) * H[n-1]
-         end
+           for n in 2:nmax
+               # recursion relation for order n:
+               H[n+1] = 2x * H[n] - 2(n-1) * H[n-1]
+           end
 
-         return H
-     end;
+           return H
+     end
 
 julia> hermite_polynomials(n) = hermite_polynomials(Int, n);
 
@@ -94,7 +94,7 @@ julia> H = hermite_polynomials(10);
 
 julia> function hermite_polynomial(n::Int)
     @assert 0 ≤ n ≤ length(H) "Not enough Hermite polynomials generated"
-    return H[n+1];
+    return H[n+1]
 end
 
 julia> hermite_polynomial(6)  # degree 6
@@ -108,22 +108,23 @@ behavior under overflow of the standard `Int64` type will not suffice. Rather, t
 be generated with `hermite_polynomials(BigInt, 100)` to ensure
 the use of arbitrary length integers.
 
+## Using a generating function
 As a second example, we describe a numerical way of obtaining the
-Hermite polynomials from their generating function: the $n$th Hermite polynomial
-corresponds to the $n$th derivative of the function $\exp(2t \, x - t^2)$.
+Hermite polynomials from their generating function: the $n$-th Hermite polynomial
+corresponds to the $n$-th derivative of the function $\exp(2t \, x - t^2)$.
 
 ```julia
-julia> 𝒢(x,t) = exp(2t * x - t^2); # generating function; 𝒢 is typed as \scrG<TAB>
+julia> 𝒢(x,t) = exp(2t * x - t^2)  # generating function; 𝒢 is typed as \scrG<TAB>
 
-julia> xn = set_variables("x", numvars=1, order=10);
+julia> xn = set_variables("x", numvars=1, order=10)
 
-julia> x = xn[1];
+julia> x = xn[1]
 
-julia> t = Taylor1([zero(x), one(x)], 10); # Taylor1{TaylorN{Float64}}
+julia> t = Taylor1([zero(x), one(x)], 10)  # Taylor1{TaylorN{Float64}}
 
-julia> gf = 𝒢(x, t); # Taylor1 expansion of 𝒢
+julia> gf = 𝒢(x, t)  # Taylor1 expansion of 𝒢
 
-julia> HH(n::Int) = derivative(n, gf); # n-th derivative of `gf`
[…]
