### 2019-03-12 98281c2
Additions to the paper (#197)
* Additions to the paper

* Add two examples related to Hermite polynomials

* Add the reference to Hermite polynomials (wikipedia)

* Fix a typo and minor changes

* Simplify Hermite recurrence code; minor changes to English

* Add Taylor integration notebook

* Changes to text of paper

* Fix error

* Add Taylor integration example to paper
