+++ paper/paper.bib
@@ -2,7 +2,8 @@
   author       = {Benet, Luis and Sanders, David P.},
   title        = {A {Julia} package for {Taylor} expansions in one or
     more independent variables},
-  howpublished = {\url{https://github.com/JuliaDiff/TaylorSeries.jl}}
+  howpublished = {\url{https://github.com/JuliaDiff/TaylorSeries.jl}},
+  doi     = {10.5281/zenodo.2557003},
 }
 
 @article{julia,
@@ -57,11 +58,10 @@
  keywords  = {IEEE 754 standard, Multiple-precision arithmetic, correct rounding, elementary function, floating-point arithmetic, portable software},
 }
 
-@misc{ValidatedNumerics.jl,
+@misc{IntervalArithmetic.jl,
   author       = {Sanders, David P. and Benet, Luis},
-  title        = {ValidatedNumerics.jl: Rigorous floating-point calculations
-      with interval arithmetic in {Julia}},
-  url = {https://github.com/JuliaIntervals/ValidatedNumerics.jl},
+  title        = {IntervalArithmetic.jl: Rigorous floating-point calculations using interval arithmetic in {Julia}},
+  url = {https://github.com/JuliaIntervals/IntervalArithmetic.jl},
   year = {2019},
 }
 
@@ -71,12 +71,14 @@
   url = {https://github.com/JuliaArbTypes/ArbFloats.jl},
   year = {2019},
 }
+
 @misc{TaylorIntegration.jl,
   author       = {P\'erez-Hern\'andez, Jorge A. and Benet, Luis},
   title        = {TaylorIntegration.jl: precise {ODE} integration using {Taylor}'s
       method in {Julia}},
   url = {https://github.com/PerezHz/TaylorIntegration.jl},
   year = {2019},
+  doi = {10.5281/zenodo.2562364},
 }
 
 @misc{TaylorModels.jl,
@@ -86,3 +88,11 @@
   url = {https://github.com/JuliaIntervals/TaylorModels.jl},
   year = {2019},
 }
+
+@misc{@HermitePols_wikipedia,
+    author = "{Wikipedia contributors}",
+    title  = "Hermite polynomials --- {Wikipedia}{,} The Free Encyclopedia",
+    year   = "{2019}",
+    url    = "https://en.wikipedia.org/w/index.php?title=Hermite_polynomials&oldid=877517833",
+    note   = "[Online; accessed 3-March-2019]",
+  }
