+++ paper/paper.md
@@ -22,37 +22,171 @@ bibliography: paper.bib
 
 # Summary
 
-`TaylorSeries.jl`
-provides a framework to use and manipulate
+The purpose of `TaylorSeries.jl` is to provide a framework to exploit
 Taylor polynomials in one and more variables
-in the [Julia programming language](https://julialang.org) [@julia]. It allows to compute
-elementary functions, e.g. `exp`, of
-polynomials (`Taylor1`- or `TaylorN`-type objects), where
-techniques of automatic differentiation are used
+in the [Julia programming language](https://julialang.org) [@julia]. In
+certain cases, it provides a primitive CAS (computer algebra system),
+which works numerically and not symbolically.
+The package allows to manipulate polynomials of a specified maximum
+degree, including power and composition, as well as series expansions
+of some elementary functions on polynomials, e.g. `exp`,
+where techniques of automatic differentiation are used
 [@Tucker:ValidatedNumerics; @HaroEtAl:ParameterizMeth]. Differentiation and
 integration are also implemented.
 
-The package allows to work with different `Number` formats
-as coefficients of the series, including complex numbers,
-the arbitrary precision `BigFloat`s [@MPFR],
-`Interval`s [@ValidatedNumerics.jl], `ArbFloat`s [@ArbFloats.jl],
-as well as `Taylor1` and `TaylorN` objects.
+Two basic immutable types are defined, `Taylor1` and `TaylorN`,
+which represent the series expansions in one or several variables,
+respectively. These are essentially vectors of coefficients,
+ordered by increasing degree. In the case of `TaylorN`, the
+coefficients are `HomogeneousPolynomials`, which in turn are vectors
+of coefficients representing all monomials with a given number of variables
+and order (total degree), in some lexicographical order.
 
-`TaylorSeries.jl` is a core component of
+Julia's parametrized type system allows the construction of Taylor series whose coefficient type is any subtype of the `Number` abstract type. Use cases include complex numbers,
+arbitrary precision `BigFloat`s [@MPFR],
+`Interval`s [@IntervalArithmetic.jl], `ArbFloat`s [@ArbFloats.jl],
+as well as `Taylor1` and `TaylorN` objects themselves.
+
+`TaylorSeries.jl` is the main component of
 [`TaylorIntegration.jl`](https://github.com/PerezHz/TaylorIntegration.jl)
 [@TaylorIntegration.jl], whose aim is to perform accurate integration
 of ODEs using the Taylor method, including jet transport techniques,
-and of
+where a small region around an initial condition is integrated.
+It is also a key component of
 [`TaylorModels.jl`](https://github.com/JuliaIntervals/TaylorModels.jl)
-[@TaylorModels.jl],
-which constructs rigorous polynomial approximations of functions.
+[@TaylorModels.jl], whose aim is to construct rigorous polynomial
+approximations of functions.
 
+# Examples
+
+We present three examples to illustrate the use of `TaylorSeries.jl`. Other
+examples, as well as a detailed user guide, can be found in the
+[documentation](http://www.juliadiff.org/TaylorSeries.jl/stable).
+
+## Hermite polynomials
+As a first example we describe how to generate the [Hermite polynomials][@HermitePols_wikipedia]
+("physicist's" version) up to a given maximum order. Firstly we directly exploit the recurrence relation satisfied by the polynomials.
+
+```julia
+julia> using TaylorSeries
+
+julia> displayBigO(false)
+
[…]
