### 2019-03-12 #1043 @lbenet
@jedbrown Sorry for the long delay. I just merged https://github.com/JuliaDiff/TaylorSeries.jl/pull/197, where we addressed the suggestions by @kyleniemeyer.

We rewrote parts of the paper to clarify the purpose of the package (a polynomial manipulator implemented numerically), briefly described the way in which it is designed, emphasized our use cases, and added three examples of its usage, which are different from those already present in the documentation. The examples consist in two ways of generating the Hermite polynomials, which are directly from the recurrence relation and semi-analytically from the generating function, and a simple implementation of Picard's iteration scheme to solve initial value problems.

I hope the paper fulfill now the requirements to be accepted.
