### 2019-03-12 de5fb93
Correcting two syntax errors in the paper
