### 2019-03-13 69b986c
Correcting two syntax errors in the paper (#201)
