+++ paper/paper.bib
@@ -89,7 +89,7 @@
   year = {2019},
 }
 
-@misc{@HermitePols_wikipedia,
+@misc{HermitePols_wikipedia,
     author = "{Wikipedia contributors}",
     title  = "Hermite polynomials --- {Wikipedia}{,} The Free Encyclopedia",
     year   = "{2019}",
