### 2019-03-13 #1043 @jedbrown
Looks like trouble with unicode monospace fonts. I didn't find a simple workaround -- the characters are evidently part of DejaVu Sans Mono, which I use in my editor, but I haven't gotten pandoc/xelatex to pick them up even when setting `monofont: DejaVu Sans Mono` in the header -- it uses the font, but doesn't find some characters that work in my editor. Perhaps @arfon has ideas.

Though the accessibility implications are tragic, embedding a PDF figure might be the simplest way to rectify this.
