### 2019-03-13 #1043 @lbenet
@jedbrown I will follow your suggestion embedding pdf (or perhaps png's ?) screen-shots of a REPL session.

With regards to the content and changes, are they ok, or do we have to do something else?
