### 2019-03-13 #1043 @jedbrown
We're ultimately creating a PDF, so PDF embeddings tend to work better, including supporting selection of text from the embedding. PNG is a raster format so you lose quality and selection.

On content, would it make sense to write a brief intro to the mathematical representation and any pitfalls such as numerical stability or range of polynomial degrees where the package would be useful? This could be just one paragraph, but it's pretty abstract right now, before jumping to examples.
