### 2019-03-13 #1043 @lbenet
Thanks for the comments.

The pdf is fine; I'll check what's better, a screen-shot from a REPL session or a pdf output from a jupyter session. Another clarification with respect to this: I guess that the pdf figure refers only to the final output of the Hermite polynomials, keeping the code as a block-code, right? Or do you prefer everything (the three examples) as pdf-figures?

Regarding the content you suggest, I'll think it over and come back to you about it. Right now my concern would be to make too specialized to a broad audience...
