### 2019-03-13 #1043 @jedbrown
A screenshot would typically be a raster, so not ideal. I'm not sure about Jupyter PDF output.
