### 2019-03-13 #1043 @lbenet
I think the pdf output of a Jupyter notebook is created with an intermediate LaTeX step; I'll test that and upload the figs to the repo so you can inspect them.
