### 2019-03-13 78dd0ec
Add Fig1.pdf and Fig2.pdf, and include them in paper.md
