### 2019-03-13 #1043 @lbenet
I have opened https://github.com/JuliaDiff/TaylorSeries.jl/pull/202, where you can find the pdf's corresponding to the output which is not properly displayed. @jedbrown Is this the solution you have in mind?

The pdf's were generated (after small modifications) from the LaTeX file obtained directly from the Jupyter notebook.
