### 2019-03-13 #1043 @jedbrown
The GitHub PDF viewer is junk, but if you download the PDF to use a real viewer, the text is selectable (as desired). Size is wonky. What font are you using in the LaTeX? Does it work for the blocks with Unicode characters? Pandoc uses LaTeX (well, XeTeX) so the font issue may be solvable that way as well.
