### 2019-03-13 #1043 @lbenet
Thanks @jedbrown for generating the article proof; I'll experiment with the `width` specification forthe figures, which hopefully generates nicer size as well as figure position (e.g., after the code).

I regarding the LaTeX file used, the following packages may be relevant for the font issues: `fontec`, `mathpazo`, `ucs` (extended unicode support), `inputenc` (allow unicode in tex document).
