### 2019-03-13 abb108c
Experiment with the width specification of figures
