### 2019-03-13 0dce5e9
Adjustement on the width of figs
