### 2019-03-13 #1043 @lbenet
@jedbrown Would the current output be acceptable? (I managed to place the figures just after the code!)
