### 2019-03-13 #1043 @jedbrown
@lbenet So long as you're getting this from Jupyter, I think it would be far less distracting to have the entire code snippet in a Jupyter cell rather than line-by-line entry with `julia>`.
