### 2019-03-13 #1043 @lbenet
Ok, I'll change it and push another commit. The reason I kept the code-block thinking on the REPL is that it allows to copy-past (to perhaps play with it). I'm not sure if having it in the pdf allows for the same.
