### 2019-03-13 #1043 @jedbrown
It does. That was part of the rationale for including those as PDF instead of PNG.
