### 2019-03-13 e306d6f
Delete code blocks, and replace them by pdf figures.
