+++ paper/paper.md
@@ -40,7 +40,12 @@ respectively. These are essentially vectors of coefficients,
 ordered by increasing degree. In the case of `TaylorN`, the
 coefficients are `HomogeneousPolynomials`, which in turn are vectors
 of coefficients representing all monomials with a given number of variables
-and order (total degree), in some lexicographical order.
+and order (total degree), in some lexicographical order. Higher degree
+of the polynomials, specially for several variables, imply more
+memory allocations; while we have not tested the limits
+of the degree of the polynomials that can be used, `TaylorSeries.jl`
+has been successfully used with 80-th degree `Taylor1` polynomials,
+as well as 60-th degree `TaylorN` polynomials in 4 variables.
 
 Julia's parametrized type system allows the construction of Taylor series whose coefficient type is any subtype of the `Number` abstract type. Use cases include complex numbers,
 arbitrary precision `BigFloat`s [@MPFR],
@@ -66,45 +71,11 @@ examples, as well as a detailed user guide, can be found in the
 ## Hermite polynomials
 As a first example we describe how to generate the [Hermite polynomials][@HermitePols_wikipedia]
 ("physicist's" version) up to a given maximum order. Firstly we directly exploit the recurrence relation satisfied by the polynomials.
-The output of the last command is displayed in Figure 1.
 
-```julia
-julia> using TaylorSeries
-
-julia> displayBigO(false)
-
-julia> function hermite_polynomials(::Type{T}, nmax::Int) where {T <: Integer}
-
-           x = Taylor1(T, nmax)    # Taylor variable
-           H = fill(x, nmax + 1)   # vector of Taylor series to be overwritten
-
-           H[1] = 1   # order 0
-           H[2] = 2x  # order 1
-
-           for n in 2:nmax
-               # recursion relation for order n:
-               H[n+1] = 2x * H[n] - 2(n-1) * H[n-1]
-           end
-
-           return H
-     end
-
-julia> hermite_polynomials(n) = hermite_polynomials(Int, n);
-
-julia> H = hermite_polynomials(10);
-
-julia> function hermite_polynomial(n::Int)
-    @assert 0 ≤ n ≤ length(H) "Not enough Hermite polynomials generated"
-    return H[n+1]
-end
-
-julia> hermite_polynomial(6)  # degree 6
-```
-
-![Output of `hermite_polynomial(6)` constructed directly from the recursion relation.](Fig1.pdf){width=130%}  
+![Code to generate Hermite polynomials directly from the recursion relation; the last line displays the 6th Hermite polynomial.](Fig1.pdf){width=130%}  
 The example above can be slightly modified to compute, for example, the 100th Hermite polynomial.
 In this case, the coefficients will be larger than $2^{63}-1$, so the modular
-behavior under overflow of the standard `Int64` type will not suffice. Rather, the polynomials should
+behavior, under overflow of the standard `Int64` type, will not suffice. Rather, the polynomials should
 be generated with `hermite_polynomials(BigInt, 100)` to ensure
 the use of arbitrary length integers.
 
@@ -113,27 +84,11 @@ As a second example, we describe a numerical way of obtaining the
 Hermite polynomials from their generating function: the $n$-th Hermite polynomial
 corresponds to the $n$-th derivative of the function $\exp(2t \, x - t^2)$.
 
-```julia
-julia> 𝒢(x,t) = exp(2t * x - t^2)  # generating function; 𝒢 is typed as \scrG<TAB>
-
-julia> xn = set_variables("x", numvars=1, order=10)
-
-julia> x = xn[1]
-
[…]
