### 2019-03-14 6297c0c
Adjust width of figures
