### 2019-03-14 #1043 @lbenet
@jedbrown I pushed a couple of commits to properly adjust the figures. I think they are ok, including the font size displayed. For consistency, the code corresponding to the last example is also added as  figure 3.

I also have added a sentence describing the limits until which we have tested the package. Those are no hard-limits, but simply usage limits that have been sufficient for us.
