### 2019-03-14 #1043 @lbenet
I just noticed that subscripts and superscripts are gone again...
