### 2019-03-14 #1043 @lbenet
... Though this seems to occur in the generated pdf; the original pdf files display them properly.
