### 2019-03-14 #1043 @dpsanders
Unicode symbols seem to be missing from the generated PDF.
