### 2019-03-14 #1043 @jedbrown
Looks like Whedon didn't listen to your instruction to use the branch (but did heed mine). Can you check that once more to confirm whether it's a bug in Whedon?

Characters look fine to me. Some of the comments run really long. Could you shorten them or put them on their own line? There's also an almost-awkward amount of vertical whitespace compared to the main text.
