### 2019-03-25 dbf27d8
Clarifications to paper
