### 2019-03-25 #1043 @lbenet
@jedbrown We have just addressed [your comments](https://github.com/openjournals/joss-reviews/issues/1043#issuecomment-472541537) on the content. The pdf looks ok to me. I am waiting for the green lights of travis to merge to master. Is there something else to be done?
