### 2019-03-25 5456402
Add pdfs to paper (#202)
* Add Fig1.pdf and Fig2.pdf, and include them in paper.md

* Experiment with the width specification of figures

* Adjustement on the width of figs

* Delete code blocks, and replace them by pdf figures.

* Crop images to adjust them

* Adjust width of figures

* Clarifications to paper

* Fix a minor typo
