+++ paper/paper.md
@@ -22,25 +22,26 @@ bibliography: paper.bib
 
 # Summary
 
-The purpose of `TaylorSeries.jl` is to provide a framework to exploit
-Taylor polynomials in one and more variables
-in the [Julia programming language](https://julialang.org) [@julia]. In
-certain cases, it provides a primitive CAS (computer algebra system),
+The purpose of the `TaylorSeries.jl` package is to provide a framework to exploit Taylor polynomials in one and several variables
+in the [Julia programming language](https://julialang.org) [@julia].
+It can be thought of as providing a primitive CAS (computer algebra system),
 which works numerically and not symbolically.
-The package allows to manipulate polynomials of a specified maximum
-degree, including power and composition, as well as series expansions
-of some elementary functions on polynomials, e.g. `exp`,
+The package allows the user to define dense polynomials $p(x)$ of one variable and $p(\mathbf{x})$ of several variables with a specified maximum degree, and perform operations on them, including powers and composition, as well as series expansions
+for elementary functions of polynomials, for example $\exp[p(x)]$,
 where techniques of automatic differentiation are used
 [@Tucker:ValidatedNumerics; @HaroEtAl:ParameterizMeth]. Differentiation and
 integration are also implemented.
 
-Two basic immutable types are defined, `Taylor1` and `TaylorN`,
-which represent the series expansions in one or several variables,
-respectively. These are essentially vectors of coefficients,
-ordered by increasing degree. In the case of `TaylorN`, the
-coefficients are `HomogeneousPolynomials`, which in turn are vectors
+Two basic immutable types are defined, `Taylor1{T}` and `TaylorN{T}`,
+which represent polynomials in one and several variables, respectively; the maximum degree is a field of the types. These types are parametrized by the type `T` of the polynomial coefficients; they essentially consist of one-dimensional arrays of coefficients, ordered by increasing degree.
+
+In the case of `TaylorN`, the
+coefficients are `HomogeneousPolynomial`s, which in turn are vectors
 of coefficients representing all monomials with a given number of variables
-and order (total degree), in some lexicographical order.
+and order (total degree), ordered lexicographically. Higher degree
+polynomials require more
+memory allocation, especially for several variables; while we have not extensively tested the limits of the degree of the polynomials that can be used, `Taylor1` polynomials up to degree 80 and
+`TaylorN` polynomials up to degree 60 in 4 variables have been successfully used. Note that the current implementation of multi-variable series builds up extensive tables in memory which allow to speed up the index calculations.
 
 Julia's parametrized type system allows the construction of Taylor series whose coefficient type is any subtype of the `Number` abstract type. Use cases include complex numbers,
 arbitrary precision `BigFloat`s [@MPFR],
@@ -67,73 +68,23 @@ examples, as well as a detailed user guide, can be found in the
 As a first example we describe how to generate the [Hermite polynomials][@HermitePols_wikipedia]
 ("physicist's" version) up to a given maximum order. Firstly we directly exploit the recurrence relation satisfied by the polynomials.
 
-```julia
-julia> using TaylorSeries
-
-julia> displayBigO(false)
-
-julia> function hermite_polynomials(::Type{T}, nmax::Int) where {T <: Integer}
-
-           x = Taylor1(T, nmax)    # Taylor variable
[…]
