### 2019-03-25 #1043 @lbenet
I just merged https://github.com/JuliaDiff/TaylorSeries.jl/pull/202
