### 2019-04-02 #1043 @jedbrown
@lbenet Thanks for your patience. (I was offline last week.) Please fix this citation https://github.com/JuliaDiff/TaylorSeries.jl/commit/98281c271e83193037c7805a2ccf2d6711c50b74#r33009220
There is also a comment in figure 3 that spills into the margin. Can you shorten it? We should be ready to archive once these formatting issues are fixed.
