+++ paper/paper.bib
@@ -88,11 +88,3 @@
   url = {https://github.com/JuliaIntervals/TaylorModels.jl},
   year = {2019},
 }
-
-@misc{HermitePols_wikipedia,
-    author = "{Wikipedia contributors}",
-    title  = "Hermite polynomials --- {Wikipedia}{,} The Free Encyclopedia",
-    year   = "{2019}",
-    url    = "https://en.wikipedia.org/w/index.php?title=Hermite_polynomials&oldid=877517833",
-    note   = "[Online; accessed 3-March-2019]",
-  }
