### 2019-04-03 29b376e
Add tex files to generate pdf's of figures
