### 2019-04-03 #1043 @lbenet
@jedbrown I just opened https://github.com/JuliaDiff/TaylorSeries.jl/pull/203; the citation was fixed (link to Wikipedia) and the comment of Fig 3 shortened. I also added the tex files to generate the pdf's (I cropped the pdf's manually though).
