### 2019-04-04 #1043 @jedbrown
Looks good to me, thanks. Please merge and tag ([annotated tag](https://git-scm.com/book/en/v2/Git-Basics-Tagging#_annotated_tags) preferred) and archive on Zenodo or similar, then report the DOI back here.
