### 2019-04-04 5804bdf
Final touches to paper (#203)
* Link to wikipedia page instead of bib reference

* Shorten a comment in Fig3

* Add tex files to generate pdf's of figures
