+++ v0.9.2/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"#TaylorSeries.jl-1","page":"Home","title":"TaylorSeries.jl","text":"","category":"section"},{"location":"#","page":"Home","title":"Home","text":"A Julia package for Taylor expansions in one or more independent variables.","category":"page"},{"location":"#","page":"Home","title":"Home","text":"","category":"page"},{"location":"#Authors-1","page":"Home","title":"Authors","text":"","category":"section"},{"location":"#","page":"Home","title":"Home","text":"Luis Benet, Instituto de Ciencias   Físicas, Universidad Nacional Autónoma de México (UNAM).\nDavid P. Sanders, Facultad   de Ciencias, Universidad Nacional Autónoma de México (UNAM).","category":"page"},{"location":"#License-1","page":"Home","title":"License","text":"","category":"section"},{"location":"#","page":"Home","title":"Home","text":"TaylorSeries is licensed under the MIT \"Expat\" license; see LICENSE for the full license text.","category":"page"},{"location":"#Installation-1","page":"Home","title":"Installation","text":"","category":"section"},{"location":"#","page":"Home","title":"Home","text":"TaylorSeries.jl is a registered package, and is simply installed by running","category":"page"},{"location":"#","page":"Home","title":"Home","text":"pkg> add(\"TaylorSeries\")","category":"page"},{"location":"#Related-packages-1","page":"Home","title":"Related packages","text":"","category":"section"},{"location":"#","page":"Home","title":"Home","text":"Polynomials.jl: Polynomial manipulations\nPowerSeries.jl: Truncated power series for Julia\nMultivariatePolynomials.jl: Multivariate polynomials interface\nAbstractAlgebra.jl: Generic abstract algebra functionality in pure Julia\nForwardDiff.jl: Forward Mode Automatic Differentiation for Julia\nReverseDiff.jl: Reverse Mode Automatic Differentiation for Julia\nHyperDualNumbers.jl: Julia implementation of HyperDualNumbers","category":"page"},{"location":"#Acknowledgments-1","page":"Home","title":"Acknowledgments","text":"","category":"section"},{"location":"#","page":"Home","title":"Home","text":"This project began (using Python) during a Masters' course in the postgraduate programs in Physics and in Mathematics at UNAM, during the second half of 2013. We thank the participants of the course for putting up with the half-baked material and contributing energy and ideas.","category":"page"},{"location":"#","page":"Home","title":"Home","text":"We acknowledge financial support from DGAPA-UNAM PAPIME grants PE-105911 and PE-107114, and PAPIIT grants IG-101113 and IG-100616. LB acknowledges support through a Cátedra Marcos Moshinsky (2013).","category":"page"},{"location":"background/#Background-1","page":"Background","title":"Background","text":"","category":"section"},{"location":"background/#","page":"Background","title":"Background","text":"","category":"page"},{"location":"background/#Introduction-1","page":"Background","title":"Introduction","text":"","category":"section"},{"location":"background/#","page":"Background","title":"Background","text":"TaylorSeries.jl is an implementation of high-order automatic differentiation, as presented in the book by W. Tucker [1]. The general idea is the following.","category":"page"},{"location":"background/#","page":"Background","title":"Background","text":"The Taylor series expansion of an analytical function f(t) with one independent variable t around t_0 can be written as","category":"page"},{"location":"background/#","page":"Background","title":"Background","text":"beginequation\nf(t) = f_0 + f_1 (t-t_0) + f_2 (t-t_0)^2 + cdots + f_k (t-t_0)^k + cdots\nendequation","category":"page"},{"location":"background/#","page":"Background","title":"
[…]
