### 2019-04-04 #1043 @lbenet
Thanks a lot @jedbrown for all your work and comments.

I merged JuliaDiff/TaylorSeries.jl#203, just tagged a new minor version (0.9.2), which has been accepted in METADATA.jl (current central repo of registered packages, but not for long). (I hope I annotated it properly; sorry if not.) The DOI at zenodo (v0.9.2) is: [10.5281/zenodo.2628898](http://doi.org/10.5281/zenodo.2628898).
